+++ RELEASE_NOTES.md
@@ -2,6 +2,11 @@
 
 ### Requirements
 
-### Version 0.10.0-prerel1
+### Version 0.10.0-prerel3
 #### Features
 * Added HTML help - use `make docs` to create it
+* Added support for weights with non-random columns
+* Added support for generation of streaming data sets
+* Added support for multiple dependent base columns
+* Added support for scripting of table create statements
+